--- src/ss/ita/Main.java
@@ -1,8 +1,0 @@
-package ss.ita;
-
-public class Main {
-
-    public static void main(String[] args) {
-	// write your code here
-    }
-}
