+++ src/ss/ita/kata/implementation/liuba/LiubaEight.java
@@ -1,8 +1,8 @@
-package ss.ita.implementation.nik;
+package ss.ita.kata.implementation.liuba;
 
 import ss.ita.kata.Eight;
 
-public class NikEight implements Eight {
+public class LiubaEight implements Eight {
     @Override
     public int Liters(double time) {
         return 0;
